requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
groq==0.4.1
openai==1.3.0
requests-oauthlib==1.3.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
twilio==8.10.0
websockets==12.0
